refactor: resource 파일 TypeScript로 변경
